Support MSIX bundles
MSIX bundles are just APPX bundles containing packages named ".msix"
instead of ".appx".

This patch treats msix files the same as appx files.

i.e. they are not added to the blockmap and they are not compressed.

Without this, opening an MSIX bundle gives the error "Unable to parse
package".