Add AppxManifest.xml after everything else
For whatever reason, the AppxManifest needs to come after the other
files in the ZIP for signing with tools like relic to work.

If you look at other example APPX or MSIX packages, they generally end
like this:

...application files...
AppxManifest.xml
AppxBlockMap.xml
[Content_Types].xml
AppxMetadata/CodeIntegrity.cat
AppxSignature.p7x

This patch just notices the manifest when scanning the files and adds it
to the ZIP after the scan and before the blockmap.